+++ frontend/components/commons/forms/questions-form/fields/shared-components/label-selection/LabelSelection.component.vue
@@ -1,6 +1,9 @@
 <template>
   <div class="container">
-    <div class="component-header" v-if="showSearch || showCollapseButton">
+    <div
+      class="component-header"
+      v-if="showSearch || showCollapseButton"
+    >
       <div class="left-header">
         <SearchLabelComponent
           ref="searchComponentRef"
@@ -62,7 +65,10 @@
         />
       </div>
     </transition-group>
-    <i class="no-result" v-if="!filteredOptions.length" />
+    <i
+      class="no-result"
+      v-if="!filteredOptions.length"
+    />
   </div>
 </template>