+++ frontend/components/page-contents/feedback-task-content/RecordFeedbackTaskAndQuestionnaire.content.vue
@@ -1,5 +1,8 @@
 <template>
-  <div v-if="!$fetchState.pending && !$fetchState.error" class="wrapper">
+  <div
+    v-if="!$fetchState.pending && !$fetchState.error"
+    class="wrapper"
+  >
     <template v-if="!!record">
       <RecordFeedbackTaskComponent
         :recordStatus="record.status"
@@ -17,7 +20,10 @@
       />
     </template>
 
-    <div v-else class="wrapper--empty">
+    <div
+      v-else
+      class="wrapper--empty"
+    >
       <p
         v-if="!records.hasRecordsToAnnotate"
         class="wrapper__text --heading3"