+++ src/argilla/client/feedback/dataset/remote.py
@@ -22,6 +22,10 @@ from argilla.client.feedback.dataset.base import FeedbackDatasetBase
 from argilla.client.feedback.schemas.records import FeedbackRecord, RemoteFeedbackRecord
 from argilla.client.sdk.users.models import UserRole
 from argilla.client.sdk.v1.datasets import api as datasets_api_v1
+<<<<<<< HEAD
+=======
+from argilla.client.sdk.v1.records import api as records_api_v1
+>>>>>>> a8abfaa6... Releases/1.14.0 (#3551)
 from argilla.client.utils import allowed_for_roles
 
 if TYPE_CHECKING:
@@ -172,13 +176,23 @@ class RemoteFeedbackRecords:
 
     def add(
         self,
+<<<<<<< HEAD
         records: List[FeedbackRecord],
+=======
+        records: Union[FeedbackRecord, Dict[str, Any], List[Union[FeedbackRecord, Dict[str, Any]]]],
+>>>>>>> a8abfaa6... Releases/1.14.0 (#3551)
         show_progress: bool = True,
     ) -> None:
         """Pushes a list of `FeedbackRecord`s to Argilla.
 
         Args:
+<<<<<<< HEAD
             records: A list of `FeedbackRecord`s to push to Argilla.
+=======
+            records: can be a single `FeedbackRecord`, a list of `FeedbackRecord`,
+                a single dictionary, or a list of dictionaries. If a dictionary is provided,
+                it will be converted to a `FeedbackRecord` internally.
+>>>>>>> a8abfaa6... Releases/1.14.0 (#3551)
             show_progress: Whether to show a `tqdm` progress bar while pushing the records.
 
         Raises:
@@ -202,6 +216,27 @@ class RemoteFeedbackRecords:
                 records=records_batch,
             )
 
+<<<<<<< HEAD
+=======
+    def delete(
+        self,
+        records: List[RemoteFeedbackRecord],
+    ) -> None:
+        """Deletes a list of `RemoteFeedbackRecord`s from Argilla.
+
+        Args:
+            records: A list of `RemoteFeedbackRecord`s to delete from Argilla.
+
+        Raises:
+            RuntimeError: If the deletion of the records from Argilla fails.
+        """
+        for record in records:
+            try:
+                records_api_v1.delete_record(client=self._dataset._client, id=record.id)
+            except Exception as e:
+                raise RuntimeError(f"Failed to delete record with id {record.id} from Argilla.") from e
+
+>>>>>>> a8abfaa6... Releases/1.14.0 (#3551)
 
 class RemoteFeedbackDataset(FeedbackDatasetBase):
     def __init__(
@@ -348,7 +383,13 @@ class RemoteFeedbackDataset(FeedbackDatasetBase):
             questions=self.questions,
             guidelines=self.guidelines,
         )
+<<<<<<< HEAD
         instance.add_records([record.dict(exclude={"client", "name2id"}) for record in self._records])
+=======
+        instance.add_records(
+            [record.dict(exclude={"client", "name2id", "id"}, exclude_none=True) for record in self._records]
+        )
+>>>>>>> a8abfaa6... Releases/1.14.0 (#3551)
         return instance
 
     def add_records(
@@ -359,8 +400,14 @@ class RemoteFeedbackDataset(FeedbackDatasetBase):
         """Adds the given records to the dataset and pushes those to Argilla.
 
         Args:
+<<<<<<< HEAD
             records: the records to add to the dataset. Can be a single record, a list
                 of records or a dictionary with the fields of the record.
+=======
+            records: can be a single `FeedbackRecord`, a list of `FeedbackRecord`,
+                a single dictionary, or a list of dictionaries. If a dictionary is provided,
+                it will be converted to a `FeedbackRecord` internally.
+>>>>>>> a8abfaa6... Releases/1.14.0 (#3551)
 
         Raises:
             ValueError: if the given records are an empty list.
@@ -371,6 +418,23 @@ class RemoteFeedbackDataset(FeedbackDatasetBase):
         self._records.add(records=records, show_progress=show_progress)
 
     @allowed_for_roles(roles=[UserRole.owner, UserRole.admin])
+<<<<<<< HEAD
+=======
+    def delete_records(self, records: Union["RemoteFeedbackRecord", List["RemoteFeedbackRecord"]]) -> None:
+        """Deletes the given records from the dataset in Argilla.
+
+        Args:
+            records: the records to delete from the dataset. Can be a single record or a list
+                of records. But those need to be previously pushed to Argilla, otherwise
+                they won't be deleted.
+
+        Raises:
+            RuntimeError: If the deletion of the records from Argilla fails.
+        """
+        self._records.delete(records=[records] if not isinstance(records, list) else records)
+
+    @allowed_for_roles(roles=[UserRole.owner, UserRole.admin])
+>>>>>>> a8abfaa6... Releases/1.14.0 (#3551)
     def delete(self) -> None:
         """Deletes the current `FeedbackDataset` from Argilla. This method is just working
         if the user has either `owner` or `admin` role.